initial commit of kms testing
