+++ .env
@@ -1,0 +1,3 @@
+BUILD_HARNESS_REPO=ghcr.io/defenseunicorns/not-a-build-harness/not-a-build-harness
+# renovate: datasource=docker depName=ghcr.io/defenseunicorns/not-a-build-harness/not-a-build-harness versioning=docker
+BUILD_HARNESS_VERSION=0.0.23
