+++ .github/ISSUE_TEMPLATE/general_issue.md
@@ -1,0 +1,48 @@
+---
+name: General Issue
+about: Suggest a new feature, report a bug, or just ask a question
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+<!-- Thank you for submitting an issue! Please answer the following questions. This template is intended to guide the issue author toward meeting our Definition of Ready. You don't have to use this exact format if you don't want to, but whatever you submit must meet the DoR or the issue may be closed, and you will be asked to resubmit. -->
+
+### Persona
+
+<!-- Who is submitting this request? Please describe how you use this repo. -->
+
+
+
+### Description
+
+<!-- What **outcome** is being requested? Note: An outcome is a description of the behavior desired rather than an imperative. For example, "I want to be able to create a new user" is an outcome, while "Create new users using this specific javascript library" is an imperative. -->
+
+
+
+### Use Case
+
+<!-- Why is this needed? Is this request related to a problem? Please describe. -->
+
+
+
+### Impact
+
+<!-- What will happen if this request is not fulfilled? -->
+
+
+
+### Completion
+
+<!-- What does being "done" look like?
+Note: Some "givens" of done criteria that don't need to be included in the answer:
+> * The change has been peer reviewed
+> * Automated tests have been written and/or updated, and are passing
+> * Documentation has been written and/or updated -->
+
+
+
+### Additional Context
+
+<!-- Any additional information / visual aids that might be helpful in understanding the request? -->
