+++ .github/actions/init-pipeline/action.yml
@@ -1,0 +1,36 @@
+name: Initialize Pipeline
+
+description: Sets up caches, pulls/saves/loads docker build harness image.
+
+runs:
+  using: composite
+  steps:
+    - name: Init gopath cache
+      uses: actions/cache@v3
+      with:
+        path: "${{ github.workspace }}/.cache/go"
+        key: "gopath|${{ hashFiles('.tool-versions') }}|${{ hashFiles('go.sum') }}"
+
+    - name: Init gobuild cache
+      uses: actions/cache@v3
+      with:
+        path: "${{ github.workspace }}/.cache/go-build"
+        key: "gobuild|${{ hashFiles('.tool-versions') }}|${{ hashFiles('go.sum') }}"
+
+    - name: Init docker cache
+      id: init-docker-cache
+      uses: actions/cache@v3
+      with:
+        path: "${{ github.workspace }}/.cache/docker"
+        key: "docker|${{ hashFiles('Makefile') }}"
+
+    - name: Docker save build harness
+      if: steps.init-docker-cache.outputs.cache-hit != 'true'
+      run: |
+        make docker-save-build-harness
+      shell: bash
+
+    - name: Load build harness
+      run: |
+        make docker-load-build-harness
+      shell: bash
