+++ .github/actions/pre-commit/action.yml
@@ -1,0 +1,52 @@
+  name: Run Pre-Commit Hooks
+
+  description: Run pre-commit hooks
+
+  inputs:
+    check-type:
+      description: The type of check to run. Valid values are "all", "common", "terraform", "golang", and "renovate"
+      required: true
+      default: all
+
+  runs:
+    using: composite
+    steps:
+      - name: Init pre-commit cache
+        uses: actions/cache@v3
+        with:
+          path: "${{ github.workspace }}/.cache/pre-commit"
+          key: "pre-commit-${{inputs.check-type}}|${{hashFiles('.tool-versions')}}|${{hashFiles('.pre-commit-config.yaml')}}"
+
+      - name: Init gopath cache
+        if: inputs.check-type == 'golang'
+        uses: actions/cache@v3
+        with:
+          path: "${{ github.workspace }}/.cache/go"
+          key: "gopath|${{hashFiles('.tool-versions')}}|${{hashFiles('go.sum')}}"
+
+      - name: Init gobuild cache
+        if: inputs.check-type == 'golang'
+        uses: actions/cache@v3
+        with:
+          path: "${{ github.workspace }}/.cache/go-build"
+          key: "gobuild|${{hashFiles('.tool-versions')}}|${{hashFiles('go.sum')}}"
+
+      - name: Init docker cache
+        id: init-docker-cache
+        uses: actions/cache@v3
+        with:
+          path: "${{ github.workspace }}/.cache/docker"
+          key: "docker|${{hashFiles('.env')}}"
+
+      - name: Docker save build harness
+        if: steps.init-docker-cache.outputs.cache-hit != 'true'
+        shell: bash
+        run: make docker-save-build-harness
+
+      - name: Load build harness
+        shell: bash
+        run: make docker-load-build-harness
+
+      - name: Run `pre-commit run -a`
+        shell: bash
+        run: "make pre-commit-${{inputs.check-type}} fix-cache-permissions"
