+++ .github/workflows/pre-commit-pr.yml
@@ -1,0 +1,49 @@
+name: pre-commit
+
+on:
+  pull_request:
+
+jobs:
+  common:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: pre-commit
+        uses: ./.github/actions/pre-commit
+        with:
+          check-type: common
+
+  terraform:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: pre-commit
+        uses: ./.github/actions/pre-commit
+        with:
+          check-type: terraform
+
+  golang:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: pre-commit
+        uses: ./.github/actions/pre-commit
+        with:
+          check-type: golang
+
+  renovate:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: pre-commit
+        uses: ./.github/actions/pre-commit
+        with:
+          check-type: renovate
