+++ .github/workflows/pre-commit-trunk.yml
@@ -1,0 +1,18 @@
+name: pre-commit
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  all:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: pre-commit
+        uses: ./.github/actions/pre-commit
+        with:
+          check-type: all
