+++ .github/workflows/test-module.yml
@@ -1,0 +1,54 @@
+name: test-component
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    types:
+      - opened
+      - reopened
+      - synchronize
+      - ready_for_review
+    branches:
+      - main
+    paths:
+      - "**.tf"
+      - "**.go"
+      - ".github/**"
+
+permissions:
+  id-token: write
+  contents: read
+
+defaults:
+  run:
+    # We need -e -o pipefail for consistency with GitHub Actions' default behavior
+    shell: bash -e -o pipefail {0}
+
+jobs:
+  # Run the tests
+  run-test:
+    name: Test Component
+    runs-on: ubuntu-latest
+    # Only run when the pull request is not in draft status
+    if: github.event.pull_request.draft == false
+    steps:
+
+      - name: Checkout the code
+        uses: actions/checkout@v3
+
+      - name: Initialize Pipeline
+        uses: ./.github/actions/init-pipeline
+
+      - name: Configure AWS Credentials
+        uses: aws-actions/configure-aws-credentials@v2
+        with:
+          role-to-assume: ${{ secrets.AWS_COMMERCIAL_ROLE_TO_ASSUME }}
+          role-session-name: ${{ github.event.client_payload.pull_request.head.sha || github.sha }}
+          aws-region: us-east-1
+          # 21600 seconds == 6 hours
+          role-duration-seconds: 21600
+
+      - name: Run live infra tests
+        run: |
+          make test
