+++ .tflint.hcl
@@ -1,0 +1,10 @@
+plugin "terraform" {
+  enabled = true
+  preset  = "recommended"
+}
+
+#plugin "aws" {
+#  enabled = true
+#  version = "0.23.0"
+#  source  = "github.com/terraform-linters/tflint-ruleset-aws"
+#}
