+++ .tool-versions
@@ -1,0 +1,20 @@
+# renovate: datasource=github-tags depName=npryce/adr-tools
+adr-tools 3.0.0
+awscli 2.11.20
+# renovate: datasource=github-tags depName=bridgecrewio/checkov
+checkov 2.3.234
+golang 1.20.4
+# renovate: datasource=github-tags depName=golangci/golangci-lint
+golangci-lint 1.52.2
+make 4.4.1
+# renovate: datasource=github-tags depName=pre-commit/pre-commit
+pre-commit 3.3.2
+terraform 1.4.6
+# renovate: datasource=github-tags depName=terraform-docs/terraform-docs
+terraform-docs 0.16.0
+# renovate: datasource=github-tags depName=terraform-linters/tflint
+tflint 0.46.1
+# renovate: datasource=github-tags depName=aquasecurity/tfsec
+tfsec 1.28.1
+# renovate: datasource=github-tags depName=defenseunicorns/zarf
+zarf 0.26.4
